feat: Added User Permissions matching capability
Added environmental variables PUID and PGID that can be used to match container user and hoser user to same ID # for easy file management.
To do This I added gosu to the installed packages and added commands to the entrypoint.
Also updated docker-compose.yaml and docker run examples to include the new environmental variables. Need to add instructions to the readme on how to use.